Update mini.ino
Added serial communication. Added the ability to write text on the LCD by inputting it into the serial termininal. Added length detection to the lcdEvent string( <16, 16-32, >32), however it doesn't do anything useful yet. Added "lcdOn" variable, to allow the LCD to be turned off. (This defaults to "1" or on in the code). Changed the default text to "Hello there", from "Good afternoon, I`m Victor". (Not sure about that one). Renamed string "event" to "lcdEvent". Goodnight. I apologise if I missed anything.